Safety first!
Added safeties preventing the timer being triggered like a million times
a second on account of ADC being crazy. Now the program looks for the
first dip below the threshold and ignores the rest for a set short time
allowing the light level to calm down and clear the threshold.
Alternative would be to take an average over time from the light sensor
which I might still do but I think that this is sufficient and much
faster allowing for greater accuracy.